+++ src/app/main/depart/depart.component.ts
@@ -43,26 +43,42 @@ export class DepartComponent {
   }
 
   deleteDepart(id :any,dateHeure :any){
-    this.departService.deleteDepart(id,dateHeure).subscribe({
-      next:(data)=>{
-        if(data.body.message =="success"){
-          this.router.navigate(["/home/depart"])
-          this.showAlertMessage("Success","Depart deleted successfully","success")
-        }
-        else{
-          console.log(data.body.message)
-          this.showAlertMessage("Error","Error when deleting Depart","error")
-        }
-  
-      },
-      error:(err) => {
-        console.log(err);
-        this.showAlertMessage("Error","Internal Server Error","danger")
-  
+    Swal.fire({
+      title: 'Do you want to delete this item?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Delete',
+      cancelButtonText: 'Cancel',
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.departService.deleteDepart(id,dateHeure).subscribe({
+          next:(data)=>{
+            if(data.body.message =="success"){
+              this.router.navigate(["/home/depart"])
+              this.showAlertMessage("Success","Depart deleted successfully","success")
+            }
+            else{
+              console.log(data.body.message)
+              this.showAlertMessage("Error","Error when deleting Depart","error")
+            }
+      
+          },
+          error:(err) => {
+            console.log(err);
+            this.showAlertMessage("Error","Internal Server Error","danger")
+      
+          }
+        })
       }
-    })
+    });
+ 
     
   }
+
+
+
   showAlertMessage( title:string, message:string, icon:any ){
     return Swal.fire({
