+++ src/app/main/passager/passager.component.ts
@@ -44,25 +44,39 @@ export class PassagerComponent {
   }
 
   deletePassager(id :any){
-    this.passagerService.deletePassager(id).subscribe({
-      next:(data)=>{
-        if(data.body.message =="success"){
-          this.showAlertMessage("Success","Passager deleted successfully","success")
-        }
-        else{
-          console.log(data.body.message)
-          this.showAlertMessage("Error","Error when deleting ","warning")
-        }
-  
-      },
-      error:(err) => {
-        console.log(err);
-        this.showAlertMessage("Error","Erreur au niveau du serveur","danger")
-  
+    Swal.fire({
+      title: 'Do you want to delete this item?',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonText: 'Delete',
+      cancelButtonText: 'Cancel',
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.passagerService.deletePassager(id).subscribe({
+          next:(data)=>{
+            if(data.body.message =="success"){
+              this.showAlertMessage("Success","Passager deleted successfully","success")
+            }
+            else{
+              console.log(data.body.message)
+              this.showAlertMessage("Error","Error when deleting ","warning")
+            }
+      
+          },
+          error:(err) => {
+            console.log(err);
+            this.showAlertMessage("Error","Internal server error","danger")
+      
+          }
+        })
       }
-    })
-    
+    });    
   }
+
+
+
   showAlertMessage( title:string, message:string, icon:any ){
     return Swal.fire({
